// src/app/layouts/shared/models/DeviceDto.ts
export class Device {
    id:number=0;
    name: string = "";
    make:string="";
    modal:string="";
    serialNo: string = "";
    is_Face:boolean=false;
    is_Bio:boolean=false;
    is_RFID:boolean=false;
    username:string="";
    mode: string = "";
    iP: string = "";
    port: number = 0;
    network: string = "";
    deviceType: string = "";
    machineno:string;
    machinepasscode:string;
    remoteurl:string;
    devicetoken:string;
    city: string = "";
    address: string = "";
    securityLevel: string = "";
    resourceType: string = "";
    password:string="";
}